+++ tasmota/tasmota_globals.h
@@ -122,11 +122,11 @@ String EthernetMacAddress(void);
 #define WS2812_LEDS                 30         // [Pixels] Number of LEDs
 #endif
 
-#ifdef USE_MQTT_TLS
-  const uint16_t WEB_LOG_SIZE = 2000;          // Max number of characters in weblog
-#else
+//#ifdef USE_MQTT_TLS                            // Set to 4000 on 20200922 per #9305
+//  const uint16_t WEB_LOG_SIZE = 2000;          // Max number of characters in weblog
+//#else
   const uint16_t WEB_LOG_SIZE = 4000;          // Max number of characters in weblog
-#endif
+//#endif
 
 #if defined(ARDUINO_ESP8266_RELEASE_2_3_0) || defined(ARDUINO_ESP8266_RELEASE_2_4_0) || defined(ARDUINO_ESP8266_RELEASE_2_4_1) || defined(ARDUINO_ESP8266_RELEASE_2_4_2) || defined(ARDUINO_ESP8266_RELEASE_2_5_0) || defined(ARDUINO_ESP8266_RELEASE_2_5_1) || defined(ARDUINO_ESP8266_RELEASE_2_5_2)
   #error "Arduino ESP8266 Core versions before 2.7.1 are not supported"